requirements.txt
typer
openai
python-dotenv
pydantic
questionary
pyperclip
